utility/requirements.txt
requests==2.31.0
psutil==5.9.5
python-dotenv==0.19.0 